--- a/julia_con/RelatedCon/src/RelatedCon.jl
+++ b/julia_con/RelatedCon/src/RelatedCon.jl
@@ -44,6 +44,7 @@ function related(posts)
     # key is every possible "tag" used in all posts
     # value is indicies of all "post"s that used this tag
     tagmap = Dict{String,Vector{UInt32}}()
+    sizehint!(tagmap, 100)
     for (idx, post) in enumerate(posts)
         for tag in post.tags
             tags = get!(() -> UInt32[], tagmap, tag)