Also removed the T, being a bit more explicit.

(the rules state that there are max 100 tags)

@Moelf