@Moelf 2023-10-29
Hmm, I don't like you're hard coding UInt32 so many times, why not just keep T?