@Moelf 2023-10-29
Also, sizehint on dictionary doesn't work the way you want it to