@Tortar 2023-10-29
I think that having T when it always fixed to be UInt32 is not useful, it is like using `T = Int` in my opinion