@Tortar 2023-10-29
I used `sizehint!` on Dictionaries sometime and it did improve performance, in this case though it is not very useful since the tags are just 100