@Moelf 2023-10-29
If we use T at least you don't have to change 10 different code when we need to change