@Tortar 2023-10-29
yes, indeed this is a good reason to keep it, will restore it then, but how that concurrent version is so slow?