@Tortar 2023-10-29
> idk, I'm gonna maybe get a Azure VM tmr...

This would probably be helpful, I have to say that I see a lot of variance on my computer as well, e.g. from the last @jinyus benchmark on 

        Benchmark 1: julia -O3 --project=Related -e "using Related; main()"
        Processing time (w/o IO): 30 milliseconds
        Processing time (w/o IO): 23 milliseconds
        Processing time (w/o IO): 22 milliseconds
        Processing time (w/o IO): 30 milliseconds
        Processing time (w/o IO): 24 milliseconds
        Processing time (w/o IO): 30 milliseconds
        Processing time (w/o IO): 30 milliseconds
        Processing time (w/o IO): 30 milliseconds # why so much oscillations?
        Processing time (w/o IO): 23 milliseconds # why so much oscillations?
        Processing time (w/o IO): 23 milliseconds
        Processing time (w/o IO): 23 milliseconds
        Processing time (w/o IO): 23 milliseconds
        Processing time (w/o IO): 23 milliseconds
          Time (mean ± σ):      3.357 s ±  0.026 s    [User: 3.208 s, System: 0.253 s]
          Range (min … max):    3.321 s …  3.393 s    10 runs