@Tortar 2023-10-29
But I see the same oscillations on my computer as well though