@Tortar 2023-10-29
Now it adds only -O3 + sizehint! which should be okay