@Moelf 2023-10-30
https://github.com/jinyus/related_post_gen/pull/182

this is known