@Moelf 2023-10-30
I'm neutral, I kinda want to remove `-O3` honestly