@jinyus 2023-10-30
main:
`Julia | 28.31 ms | 337.00 ms | 2.92 s`

This PR:
```
Julia:

        Benchmark 1: julia --project=Related -e "using Related; main()"
        Processing time (w/o IO): 23 milliseconds
        Processing time (w/o IO): 31 milliseconds
        Processing time (w/o IO): 31 milliseconds
        Processing time (w/o IO): 31 milliseconds
        Processing time (w/o IO): 24 milliseconds
        Processing time (w/o IO): 31 milliseconds
        Processing time (w/o IO): 25 milliseconds
        Processing time (w/o IO): 24 milliseconds
        Processing time (w/o IO): 31 milliseconds
        Processing time (w/o IO): 31 milliseconds
        Processing time (w/o IO): 31 milliseconds
        Processing time (w/o IO): 23 milliseconds
        Processing time (w/o IO): 30 milliseconds
          Time (mean ± σ):      3.230 s ±  0.026 s    [User: 3.090 s, System: 0.243 s]
          Range (min … max):    3.194 s …  3.268 s    10 runs
         
Julia:

        Benchmark 1: julia --project=Related -e "using Related; main()"
        Processing time (w/o IO): 481 milliseconds
        Processing time (w/o IO): 363 milliseconds
        Processing time (w/o IO): 363 milliseconds
          Time (mean ± σ):      3.727 s ±  0.025 s    [User: 3.516 s, System: 0.312 s]
          Range (min … max):    3.709 s …  3.744 s    2 runs
         
Julia:

        Benchmark 1: julia --project=Related -e "using Related; main()"
        Processing time (w/o IO): 4221 milliseconds
        Processing time (w/o IO): 3162 milliseconds
        Processing time (w/o IO): 3164 milliseconds
          Time (mean ± σ):      6.839 s ±  0.001 s    [User: 6.605 s, System: 0.329 s]
          Range (min … max):    6.838 s …  6.840 s    2 runs
```