@Moelf 2023-10-31
for me without `sizehint!()` is actually faster (Azure VM + Docker), but I'm sure it's just fluctuation, I still support removing O3