@Moelf 2023-10-31
@jinyus maybe it's time to consider use `minimum` instead of `mean` as the proxy metric because Virtual CPUs are so unstable.
https://blog.kevmod.com/2016/06/10/benchmarking-minimum-vs-average/